Rename/reformat existing testing infrastructure
