--- testing/colte_tests/nodes/command_exception.py
@@ -1,6 +1,0 @@
-class CommandException(Exception):
-  """
-  Exception representing an unexpected result or failure
-  while exectuing a command on a node/VM.
-  """
-  pass
